backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
google-cloud-vision==3.4.5

# Excel/CSV Processing
openpyxl==3.1.2
pandas==2.1.3
xlrd==2.0.1

# PDF Processing
PyPDF2==3.0.1
reportlab==4.0.7

# Email
python-dotenv==1.0.0
aiosmtplib==3.0.1
jinja2==3.1.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-slugify==8.0.1
qrcode[pil]==7.4.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS
fastapi-cors==0.0.6

# Redis (opcional)
redis==5.0.1

# Logging
loguru==0.7.2
